detect already-claimed cube
I want to keep the score in the database for easy retrieval.  Then it's
also simpler if I can update the score incrementally.  So I will need to
look at whether the cube is already claimed or not.

I don't think we specified whether stealing cubes was allowed, but it's
simpler to forbid it for now.