Fix placeholder for response on Windows Phone 8.1
On WP8.1 the text message was not line-breaking for some reason.
(.textContent has a warning about not causing a reflow,
 but .innerText didn't seem to work any better).
It cannot be reproduced in Firefox.

alert() is better.  It is able to line-break compact JSON.

Also replacing the entire page is not expected by users,
they will expect the back button to work differently than it does.