LAST-MINUTE hack bugfix for text overflow/overlap on Windows Phone
a) I'm suspicious about `height: 100%` for the pages of text.
b) Windows Phone has a difference in behaviour, where the text pages are
   causing the page to
    1) become wider than intended
    2) text is cut off on the left hand side

limit the flexbox & `height: 100%` to the fullscreen app where it is needed
(list of buttons) -  excluding it from the pages which include long
paragraphs of text.

It seems very likely this is a Windows Phone bug! (with flexbox??)